backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
pyyaml>=6.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
apscheduler>=3.10.4
aiofiles>=23.2.1
python-dotenv>=1.0.0
